+++ dropauth/account.php
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<?php
+include("./authentication.php");
+?>
+<html lang="en">
+    <head>
+        <title>DropAuth - Account</title>
+        <link href="./stylesheets/styles.css" rel="stylesheet">
+
+        <meta name="viewport" content="width=device-width, initial-scale=1">
+    </head>
+    <body>
+        <a class="button" href="../store/index.php">Return to Bubble</a>
+    </body>
+</html>