+++ dropauth/index.php
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en">
+    <head>
+        <title>DropAuth</title>
+        <link href="./stylesheets/styles.css" rel="stylesheet">
+
+        <meta name="viewport" content="width=device-width, initial-scale=1">
+    </head>
+    <body>
+        <h1>DropAuth</h1>
+        <a class="button" href="./signup.php">Sign Up</a>
+        <a class="button" href="./signin.php">Sign In</a>
+    </body>
+</html>