+++ store/failsafe.php
@@ -11,6 +11,7 @@ if (strlen($xPub) <= 10) {
     echo "</a></p>";
     exit();
 }
+
 if (!is_writable("./storedatabase.txt")) {
     echo "<p style='margin:15%;color:red;'>Error: The store database file (storedatabase.txt) is not writable to Apache/PHP! You may want to contact customer support to make the store owner aware of this issue: <a href='mailto:";
     echo $support_email;
@@ -19,13 +20,27 @@ if (!is_writable("./storedatabase.txt")) {
     echo "</a></p>";
 }
 
-if ($username == "testuser") {
-    echo "<p style='margin:15%;color:yellow;'>Notice: The current username is set to 'testuser'. This is the username of the the user Bubble comes built in with for sake of testing. If you're seeing this on a production instance of Bubble, it's highly likely that the developer forgot to set up their authentication system in 'bubble/store/authentication.php' You may want to contact customer support to make the store owner aware of this issue: <a href='mailto:";
+if (!is_writable("./productsdatabase.txt")) {
+    echo "<p style='margin:15%;color:red;'>Error: The product database file (productdatabase.txt) is not writable to Apache/PHP! You may want to contact customer support to make the store owner aware of this issue: <a href='mailto:";
     echo $support_email;
     echo "'>";
     echo $support_email;
     echo "</a></p>";
 }
+
+if (!is_writable("./configurationdatabase.txt")) {
+    echo "<p style='margin:15%;color:red;'>Error: The configuration database file (configurationdatabase.txt) is not writable to Apache/PHP! You may want to contact customer support to make the store owner aware of this issue: <a href='mailto:";
+    echo $support_email;
+    echo "'>";
+    echo $support_email;
+    echo "</a></p>";
+}
+
+$authenticationArray = unserialize(file_get_contents('../dropauth/accountDatabase.txt'));
+if (isset($authenticationArray[$admin_account]) == false) {
+    echo "<p style='margin:15%;color:yellow;'>Notice: The current admin account set in the configuration is named '" . $admin_account . "'. However, no such account exists in the authentication database. This means that anyone can sign up for an account with this name, and have full control over this Bubble instance. If you are the owner of this Bubble instance, you should sign up for an account with this name so that you can access admin controls and prevent others from being able to take over.</p>";
+}
+
 
 if ($support_email == "support@server.com") {
     echo "<p style='margin:15%;color:yellow;'>Notice: The support email defined in the configuration section is still set to the default email, " . $support_email . ". This means that you won't be able to contact the owner of this store. This email should be changed before this store is published.</p>";